fix(tests): fix JWT and rating conversion tests
- Increase token expiration time to 500ms with 100ms wait for more reliable timing
- Fix rating type assertion (float64 instead of float32)

+++ api/v1/bookstore/bookstore_converter_test.go
@@ -338,7 +338,7 @@ func TestRatingConversion(t *testing.T) {
 			}
 
 			bookDTO := ToBookDTO(bookModel)
-			assert.Equal(t, tc.rating, bookDTO.Rating)
+			assert.Equal(t, float64(tc.rating), bookDTO.Rating)
 
 			// DTO → Model
 			bookDTO2 := &dto.BookDTO{
